# SpringBoot/Documentation/OpenAPI.yaml
swagger: "2.0"
info:
  description: "Esse projeto tem como objetivo demonstrar na prática os conteúdos estudados pelo Yan Zingra Pereira"
  version: "0.0.1"
  title: "Lerning Project"
  contact:
    email: "yan.zingra12@gmail.com"

host: "127.0.0.1"
basePath: "/user"


paths:
  /user:
    get:
      tags:
        - "user"
      summary: "Return users"
      description: "Return a array of users"
      produces:
        - "application/json"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/UserList"
        "500":
          description: "Internal Server Error"
        "204":
          description: "Nenhum usuário foi encontrado"
        "400":
          description: "bad request"
    post:
      tags:
        - "user"
      summary: "Create users"
      description: "Return a array of users"
      consumes:
        - application/json
      parameters:
        - in: body
          name: "user"
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        "201":
          description: "User has been created"
        "500":
          description: "Internal Server Error"
        "204":
          description: "Nenhum usuário foi encontrado"
        "400":
          description: "Bad Request"


  /user/{id}:
    get:
      tags:
        - "user"
      summary: "Return specifc user by id"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID do usuário que deseja resgatar as informações"
          required: true
          type: "integer"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Bad request"
        "204":
          description: "id not found"
        "500":
          description: "Internal Server error"
    put:
      tags:
        - "user"
      summary: "Modify specifc user by id"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID do usuário que deseja resgatar as informações"
          required: true
          type: "integer"
      responses:
        "200":
          description: "Ok"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Bad request"
        "204":
          description: "id not found"
        "500":
          description: "Internal Server error"

    delete:
      tags:
        - "user"
      summary: "Delete user by ID"
      description: "Delete a user with ID"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the order that needs to be deleted"
          required: true
          type: "integer"
      responses:
        "200":
          description: "ok"
        "403":
          description: "Invalid ID"
        "500":
          description: "Internal Server error"

definitions:

  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      userStatus:
        type: "integer"
        description: "Descreve a atual situação do usuário"
    xml:
      name: "User"

  UserList:
    type: "array"
    items:
      $ref: "#/definitions/User"